chore: cleanup external.inc.sh

+++ resources/external.inc.sh
@@ -1,30 +1,15 @@
-#!/usr/bin/env bash
-
 #----------------------------------------------------------------------------------------
-# Source keyboard from external repository or binary download location
+# Source keyboard or model from external repository or binary download location
+# This file corresponds very closely to external.sh in keymanapp/lexical-models
 #----------------------------------------------------------------------------------------
-
-is_external_source_keyboard() {
-  # https://github.com/(owner)/(repo)/tree/(commit)/(path?)
-  [[ $1 =~ ^https:\/\/github.com\/([^\/]+)\/([^\/]+)\/tree\/([a-z0-9]{40})(\/.+)?$ ]]
-}
-
-urldecode() { echo -e "${*//%/\\x}"; } # https://stackoverflow.com/a/37840948/1836776, no +
 
 retrieve_external_keyboard() {
   # Assume we are starting in the correct folder
   [ -f external_source ] || die "No external_source file found"
 
-  local source=`urldecode $(<external_source)`
+  local source=`_urldecode $(<external_source)`
 
-  if [[ -z "$FLAG_CLEAN" ]]; then
-    verify_external_target_folder_is_clean
-  else
-    clean_external_target_folder
-    return 0
-  fi
-
-  if is_external_source_keyboard "$source"; then
+  if _is_external_source_keyboard "$source"; then
     # If the external_source file has a single line referencing a GitHub commit
     # hash, then we can treat this as a source reference.
     local repo_owner=${BASH_REMATCH[1]}
@@ -33,37 +18,38 @@ retrieve_external_keyboard() {
     local repo_path=${BASH_REMATCH[4]:-}
     [[ $repo_path =~ \.\. ]] && die "cannot contain .. in path"
 
-    retrieve_external_source_keyboard "$repo_owner" "$repo_name" "$repo_hash" "$repo_path"
+    _retrieve_external_source_keyboard "$repo_owner" "$repo_name" "$repo_hash" "$repo_path"
   else
     # Otherwise, if the external_source file consists exclusively of
     # filename=url pairs, then we treat this as a binary reference, and download
     # these files into the source folder, and create the local .source_is_binary
     # file
-    retrieve_external_binary_keyboard
-  fi
-}
-
-verify_external_target_folder_is_clean() {
-  # Remove all files except:
-  # .gitignore
-  # external_source
-  # README_EXTERNAL.md
-  local files=`find * \! -name '.gitignore' -a \! -name '.source_is_binary' -a \! -name 'external_source' -a \! -name 'README_EXTERNAL.md' -print`
-  if [ ! -z "$files" ]; then
-    echo "The target folder contains unexpected files:"
-    echo $files
-    die "Aborting build"
+    _retrieve_external_binary_keyboard
   fi
 }
 
 clean_external_target_folder() {
+  # Assume we are starting in the correct folder
+  [ -f external_source ] || die "No external_source file found"
+
   local files=`find * \! -name '.gitignore' -a \! -name 'external_source' -a \! -name 'README_EXTERNAL.md' -print`
+  rm -f .source_is_binary
   if [ ! -z "$files" ]; then
     rm -rf $files
   fi
 }
 
-retrieve_external_source_keyboard() {
+_is_external_source_keyboard() {
+  # https://github.com/(owner)/(repo)/tree/(commit)/(path?)
+  [[ $1 =~ ^https:\/\/github.com\/([^\/]+)\/([^\/]+)\/tree\/([a-z0-9]{40})(\/.+)?$ ]]
+}
+
+_urldecode() {
+  # https://stackoverflow.com/a/37840948/1836776, no +
+  echo -e "${*//%/\\x}";
+}
+
+_retrieve_external_source_keyboard() {
   local repo_owner="$1"
   local repo_name="$2"
   local repo_hash="$3"
@@ -103,9 +89,10 @@ retrieve_external_source_keyboard() {
 #
 # Download all the files referenced in external_source
 #
-retrieve_external_binary_keyboard() {
+_retrieve_external_binary_keyboard() {
   # We'll read .source line by line...
   touch .source_is_binary
+  mkdir -p source
 
   # for each line in the external_source file
   while IFS="" read -r line
@@ -118,8 +105,10 @@ retrieve_external_binary_keyboard() {
     local sha256="${BASH_REMATCH[3]}"
     [[ $filename =~ \.\. ]] && die "path cannot contain .."
     [[ $filename =~ ^/ ]] && die "path cannot start with /"
-    curl -s -L "$url" --output "$filename" --create-dirs || die "Unable to download $filename"
-    echo "$sha256 $filename" | sha256sum -c --quiet || die "Invalid checksum for $filename"
+    local path=
+    [[ ! $filename =~ .keyboard_info$ ]] && path=source/
+    curl -s -L "$url" --output "$path$filename" --create-dirs || die "Unable to download $filename"
+    echo "$sha256 $path$filename" | sha256sum -c --quiet || die "Invalid checksum for $filename"
   done < external_source
 }
